pyproject.toml
[project]
name = "graphchat"
version = "0.1.0"
description = "지능적인 주제 분기와 컨텍스트 관리를 통한 AI 대화 향상 시스템"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "falkordb>=1.0.0",
    "redis>=5.0.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.24.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.0",
    "python-dotenv>=1.0.0",
    "dependency-injector>=4.41.0",
    "google-genai>=1.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-redis",
    "types-python-jose",
]


[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

